Function special::inv_inc_beta

pub fn inv_inc_beta(alpha: f64, p: f64, q: f64, ln_beta: f64) -> f64

Compute the inverse of the incomplete beta function.

The code is based on a C implementation by John Burkardt. The original algorithm was published in Applied Statistics and is known as Algorithm AS 64 and Algorithm AS 109.