Trait special::Beta

pub trait Beta {
    fn inc_beta(self, p: Self, q: Self, ln_beta: Self) -> Self;

    fn inv_inc_beta(self, p: Self, q: Self, ln_beta: Self) -> Self;

    fn ln_beta(self, other: Self) -> Self;
}

Beta functions.

Required methods

fn inc_beta(self, p: Self, q: Self, ln_beta: Self) -> Self

Compute the regularized incomplete beta function.

The code is based on a C implementation by John Burkardt. The original algorithm was published in Applied Statistics and is known as Algorithm AS 63 and Algorithm AS 109.

fn inv_inc_beta(self, p: Self, q: Self, ln_beta: Self) -> Self

Compute the inverse of the regularized incomplete beta function.

The code is based on a C implementation by John Burkardt. The original algorithm was published in Applied Statistics and is known as Algorithm AS 64 and Algorithm AS 109.

fn ln_beta(self, other: Self) -> Self

Compute the natural logarithm of the beta function.

Loading content...